safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/stacktraces/csp-inline-warning-contains-stacktrace-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'".

This test injects an inline script from JavaScript. The resulting console error should contain a stack trace.

Message[0]: [native code]:0 Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'".

Stack Trace:

    url: [native code]
    function: appendChild
    line: 0
